Fix test

+++ src/as_execution/env.rs
@@ -92,8 +92,9 @@ pub(crate) fn get_remaining_points(
     } else {
         match env.get_exhausted_points().as_ref() {
             Some(exhausted_points) => match exhausted_points.get(store).try_into() {
-                Ok::<u64, _>(exhausted) if exhausted > 0 => return Ok(0),
-                Ok::<u64, _>(_) => (),
+                // Using i32 here because it's the type used internally by wasmer for exhausted.
+                Ok::<i32, _>(exhausted) if exhausted > 0 => return Ok(0),
+                Ok::<i32, _>(_) => (),
                 Err(_) => abi_bail!("exhausted_points has wrong type"),
             },
             None => abi_bail!("Lost reference to exhausted_points"),
